move plotContour to generic
